// src/app/modules/dashboard/models/btth.type.ts
export type SecurityEventType = 'MALWARE' | 'HUNTING' | 'INTERNET';

export type TCTTTargetType =
  | 'BOT'
  | 'Giáo xứ'
  | 'Khu công nghiệp'
  | 'Lãnh đạo'
  | 'Date'
  | 'Area';

export type MapArea = 'all' | 'north' | 'central' | 'south';

export type LayerTypes = 'base' | 'network' | 'cable' | 'alert';

export type EndpointType = 'CLIENT' | 'SERVER' | 'OTHER';

export type DeviceType = 'ROUTER' | 'SWITCH' | 'FIREWALL';

export type HotSpotType = 'IndustrialPark' | 'BOT' | 'Parish';